requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bidict==0.23.1
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
deprecation==2.1.0
distlib==0.3.8
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.3
gotrue==2.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
identify==2.5.35
idna==3.6
iniconfig==2.0.0
isort==5.13.2
netifaces==0.10.6
nodeenv==1.8.0
openai==1.14.2
packaging==24.0
pillow==10.2.0
pillow_heif==0.16.0
platformdirs==4.2.0
pluggy==1.4.0
postgrest==0.16.2
pre-commit==3.7.0
pydantic==2.6.4
pydantic_core==2.16.3
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-multipart==0.0.9
python-socketio==5.11.2
PyYAML==6.0.1
realtime==1.0.2
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sse-starlette==2.1.0
starlette==0.36.3
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.0
supafunc==0.3.3
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.25.1
websockets==11.0.3
wsproto==1.2.0
